requirements.txt
fastapi
uvicorn
chromadb
anthropic
sentence-transformers
python-multipart
pydantic
python-dotenv
